test(script): convert duplicate codes into functions to simplify them
